refactor: 하드 코딩된 값을 creator를 사용하도록 변경
