+++ src/main/java/com/bookspot/batch/stock/StockCsvMetadataCreator.java
@@ -1,14 +1,14 @@
-package com.bookspot.batch.stock.writer.file;
+package com.bookspot.batch.stock;
 
 import com.bookspot.batch.BookSpotFileConst;
 import com.bookspot.batch.global.file.FileFormat;
 import com.bookspot.batch.global.file.FileMetadata;
 
-class StockCsvMetadataCreator {
+public class StockCsvMetadataCreator {
 
     private static final String DIRECTORY_NAME = BookSpotFileConst.ROOT_DIRECTORY.concat("/stock");
 
-    protected static FileMetadata create(String libraryCode) {
+    public static FileMetadata create(String libraryCode) {
         return new FileMetadata(libraryCode, DIRECTORY_NAME, FileFormat.CSV);
     }
 }
