+++ src/main/java/com/bookspot/batch/stock/writer/file/StockFileDownloader.java
@@ -1,6 +1,7 @@
 package com.bookspot.batch.stock.writer.file;
 
 import com.bookspot.batch.global.file.NaruFileDownloader;
+import com.bookspot.batch.stock.StockCsvMetadataCreator;
 import com.bookspot.batch.stock.data.StockFileData;
 import lombok.RequiredArgsConstructor;
 import org.springframework.stereotype.Service;
